[trello-16]: BIG commit, renamed form ampc to apmc
